nlp_hub/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nlp-hub"
version = "1.0.0"
description = "Comprehensive NLP Pipeline with Chatbot, Intent Recognition, Entity Extraction, RAG, Speech, LLM, and Translation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Yacine-ai-tech", email = "siddoyacinetech227@gmail.com"}
]
keywords = ["nlp", "chatbot", "intent", "entity", "rag", "speech", "llm", "translation"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "python-dotenv>=0.21.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]
gpu = [
    "torch[cuda11]>=2.0.0",
    "faiss-gpu>=1.7.4",
]
speech = [
    "google-cloud-speech>=2.21.0",
    "google-cloud-texttospeech>=2.14.0",
]
llm = [
    "openai>=1.3.0",
    "anthropic>=0.7.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/nlp-hub"
Documentation = "https://github.com/yourusername/nlp-hub/wiki"
Repository = "https://github.com/yourusername/nlp-hub.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=src --cov-report=html"
